Fix broken HRM on the apps page (#2794)
* Split context to the separate file

* Split context to the separate file
