+++ src/apps/components/ExternalAppContext/context.ts
@@ -1,0 +1,19 @@
+import { AppDetailsUrlMountQueryParams } from "@saleor/apps/urls";
+import { AppExtensionTargetEnum } from "@saleor/graphql";
+import React from "react";
+
+export interface AppData {
+  id: string;
+  appToken: string;
+  src: string;
+  label: string;
+  target: AppExtensionTargetEnum;
+  params?: AppDetailsUrlMountQueryParams;
+}
+
+export const ExternalAppContext = React.createContext<{
+  open: boolean;
+  appData: AppData | undefined;
+  setOpen: React.Dispatch<React.SetStateAction<boolean>>;
+  setAppData: React.Dispatch<React.SetStateAction<AppData | undefined>>;
+}>(undefined);
